Add Retry middleware. Kill NetworkEvents
