+++ src/libs/Middleware/Retry.js
@@ -1,0 +1,62 @@
+import lodashGet from 'lodash/get';
+import RetryCounter from '../RetryCounter';
+import * as PersistedRequests from '../actions/PersistedRequests';
+import * as Network from '../Network';
+import Log from '../Log';
+import CONST from '../../CONST';
+
+// Keep track of retries for any non-persisted requests
+const mainQueueRetryCounter = new RetryCounter();
+
+/**
+ * @param {Object} queuedRequest
+ * @param {*} error
+ * @returns {Boolean} true if we were able to retry
+ */
+function retryFailedRequest(queuedRequest, error) {
+    // When the request did not reach its destination add it back the queue to be retried if we can
+    const shouldRetry = lodashGet(queuedRequest, 'data.shouldRetry');
+    if (!shouldRetry) {
+        return false;
+    }
+
+    const retryCount = mainQueueRetryCounter.incrementRetries(queuedRequest);
+    Log.info('[Network] A retryable request failed', false, {
+        retryCount,
+        command: queuedRequest.command,
+        error: error.message,
+    });
+
+    if (retryCount < CONST.NETWORK.MAX_REQUEST_RETRIES) {
+        Network.pushToMainQueue(queuedRequest);
+        return true;
+    }
+
+    Log.info('[Network] Request was retried too many times with no success. No more retries left');
+    return false;
+}
+
+export default (response, request, isFromSequentialQueue) => response
+    .catch((error) => {
+        if (isFromSequentialQueue) {
+            const retryCount = PersistedRequests.incrementRetries(request);
+            Log.info('Persisted request failed', false, {retryCount, command: request.command, error: error.message});
+            if (retryCount >= CONST.NETWORK.MAX_REQUEST_RETRIES) {
+                Log.info('Request failed too many times, removing from storage', false, {retryCount, command: request.command, error: error.message});
+                PersistedRequests.remove(request);
+            }
+            return;
+        }
+
+        if (retryFailedRequest(request, error)) {
+            return;
+        }
+
+        if (request.command !== 'Log') {
+            Log.hmmm('[Network] Handled error when making request', error);
+        } else {
+            console.debug('[Network] There was an error in the Log API command, unable to log to server!', error);
+        }
+
+        request.reject(new Error(CONST.ERROR.API_OFFLINE));
+    });
