+++ src/libs/Middleware/index.js
@@ -1,9 +1,11 @@
 import Logging from './Logging';
 import Reauthentication from './Reauthentication';
 import Recheck from './Recheck';
+import Retry from './Retry';
 
 export {
     Logging,
     Reauthentication,
     Recheck,
+    Retry,
 };
