+++ src/libs/Network/index.js
@@ -4,18 +4,13 @@ import HttpUtils from '../HttpUtils';
 import * as ActiveClientManager from '../ActiveClientManager';
 import CONST from '../../CONST';
 import * as PersistedRequests from '../actions/PersistedRequests';
-import RetryCounter from '../RetryCounter';
 import * as NetworkStore from './NetworkStore';
-import * as NetworkEvents from './NetworkEvents';
 import * as SequentialQueue from './SequentialQueue';
 import * as Request from '../Request';
 import {version} from '../../../package.json';
 
 // Queue for network requests so we don't lose actions done by the user while offline
 let networkRequestQueue = [];
-
-// Keep track of retries for any non-persisted requests
-const mainQueueRetryCounter = new RetryCounter();
 
 /**
  * Checks to see if a request can be made.
@@ -37,34 +32,6 @@ function canMakeRequest(request) {
     // Some requests are always made even when we are in the process of authenticating (typically because they require no authToken e.g. Log, GetAccountStatus)
     // However, if we are in the process of authenticating we always want to queue requests until we are no longer authenticating.
     return request.data.forceNetworkRequest === true || (!NetworkStore.getIsAuthenticating() && !SequentialQueue.isRunning());
-}
-
-/**
- * @param {Object} queuedRequest
- * @param {*} error
- * @returns {Boolean} true if we were able to retry
- */
-function retryFailedRequest(queuedRequest, error) {
-    // When the request did not reach its destination add it back the queue to be retried if we can
-    const shouldRetry = lodashGet(queuedRequest, 'data.shouldRetry');
-    if (!shouldRetry) {
-        return false;
-    }
-
-    const retryCount = mainQueueRetryCounter.incrementRetries(queuedRequest);
-    NetworkEvents.getLogger().info('[Network] A retryable request failed', false, {
-        retryCount,
-        command: queuedRequest.command,
-        error: error.message,
-    });
-
-    if (retryCount < CONST.NETWORK.MAX_REQUEST_RETRIES) {
-        networkRequestQueue.push(queuedRequest);
-        return true;
-    }
-
-    NetworkEvents.getLogger().info('[Network] Request was retried too many times with no success. No more retries left');
-    return false;
 }
 
 /**
@@ -92,8 +59,16 @@ function removeAllPersistableRequestsFromMainQueue() {
 /**
  * @param {Object} request
  */
-function replayRequest(request) {
+function pushToMainQueue(request) {
     networkRequestQueue.push(request);
+}
+
+/**
+ * @param {Object} request
+ */
+function replayRequest(request) {
+    pushToMainQueue(request);
+
     // eslint-disable-next-line no-use-before-define
     processNetworkRequestQueue();
 }
@@ -134,20 +109,7 @@ function processNetworkRequestQueue() {
             return;
         }
 
-        Request.process(queuedRequest)
-            .catch((error) => {
-                if (retryFailedRequest(queuedRequest, error)) {
-                    return;
-                }
-
-                if (queuedRequest.command !== 'Log') {
-                    NetworkEvents.getLogger().hmmm('[Network] Handled error when making request', error);
-                } else {
-                    console.debug('[Network] There was an error in the Log API command, unable to log to server!', error);
-                }
-
-                queuedRequest.reject(new Error(CONST.ERROR.API_OFFLINE));
-            });
+        Request.process(queuedRequest);
     });
 
     // We clear the request queue at the end by setting the queue to requestsToProcessOnNextRun which will either have some
@@ -220,4 +182,5 @@ export {
     post,
     clearRequestQueue,
     replayRequest,
+    pushToMainQueue,
 };
